+++ lib/decryption.rb
@@ -1,7 +1,11 @@
 require 'date'
 require './lib/characters_list'
+require './lib/shiftables'
+require './lib/rotatables'
 
 class Decryption
+  include Shiftable
+  include Rotatable
   attr_reader :characters, :date, :key
 
   def initialize
@@ -10,63 +14,10 @@ class Decryption
     @characters = CharactersList.new.characters
   end
 
-  def key_shift
-    key_array = key.split("")
-    sum = key_array.map do |number|
-      number.to_i
-    end
-    key_shift = sum.each_cons(2).map do |sub_array| #["0, 2"] ["2"]
-      sub_array.join
-    end
-    key_shift.map do |number|
-      number.to_i
-    end
-  end
-
-  def date_shift
-    date_shifted = (date.to_i ** 2).to_s
-    last_four_digits = date_shifted[-4..-1].split("")
-    last_four_digits.map do |number|
-      number.to_i
-    end
-  end
-
-  def total_shift
-    key_date = []
-    key_date << key_shift
-    key_date << date_shift
-    key_date.transpose.map do |sub_array|
-      sub_array.sum
-    end
-  end
-
-  def decryption_shift
+  def shift
     total_shift.map do |shift|
       -shift
     end
-  end
-
-  def a_rotation(letter)
-    a_rotation = @characters.rotate(decryption_shift[0])
-    Hash[characters.zip(a_rotation)][letter]
-  end
-
-
-  def b_rotation(letter)
-    b_rotation = @characters.rotate(decryption_shift[1])
-    Hash[characters.zip(b_rotation)][letter]
-  end
-
-
-  def c_rotation(letter)
-    c_rotation = @characters.rotate(decryption_shift[2])
-    Hash[characters.zip(c_rotation)][letter]
-  end
-
-
-  def d_rotation(letter)
-    d_rotation = @characters.rotate(decryption_shift[3])
-    Hash[characters.zip(d_rotation)][letter]
   end
 
   def decrypt(message)
