+++ lib/encryption.rb
@@ -1,7 +1,11 @@
 require 'date'
 require './lib/characters_list'
+require './lib/shiftables'
+require './lib/rotatables'
 
 class Encryption
+  include Shiftable
+  include Rotatable
   attr_reader :characters, :date, :key
 
   def initialize
@@ -14,58 +18,8 @@ class Encryption
     rand(9 ** 5).to_s.rjust(5, '0')
   end
 
-  def key_shift
-    key_array = key.split("")
-    # require 'Pry';binding.pry
-    sum = key_array.map do |number|
-      number.to_i
-    end
-    key_shift = sum.each_cons(2).map do |sub_array|
-      sub_array.join
-    end
-    key_shift.map do |number|
-      number.to_i
-    end
-  end
-
-  def date_shift
-    date_shifted = (@date.to_i ** 2).to_s
-    last_four_digits = date_shifted[-4..-1].split("")
-    last_four_digits.map do |number|
-      number.to_i
-    end
-  end
-
-  def total_shift
-    key_date = []
-    key_date << key_shift
-    key_date << date_shift
-    key_date.transpose.map do |sub_array|
-      sub_array.sum
-    end
-  end
-
-  def a_rotation(letter)
-    a_rotation = @characters.rotate(total_shift[0])
-    Hash[characters.zip(a_rotation)][letter]
-  end
-
-
-  def b_rotation(letter)
-    b_rotation = @characters.rotate(total_shift[1])
-    Hash[characters.zip(b_rotation)][letter]
-  end
-
-
-  def c_rotation(letter)
-    c_rotation = @characters.rotate(total_shift[2])
-    Hash[characters.zip(c_rotation)][letter]
-  end
-
-
-  def d_rotation(letter)
-    d_rotation = @characters.rotate(total_shift[3])
-    Hash[characters.zip(d_rotation)][letter]
+  def shift
+    total_shift
   end
 
   def encrypt(message)
