--- test/shift_gen_test.rb
@@ -1,13 +1,0 @@
-require 'minitest/autorun'
-require 'minitest/pride'
-require 'Pry'
-require 'mocha/minitest'
-require 'date'
-require './lib/characters_list'
-
-class ShiftGen < Minitest::Test
-  def test_it_exists_and_has_attributes
-    shift_gen = Shiftgen.new
-    
-  end
-end
